pyproject.toml
[tool.poetry]
name = "jjalbot"
version = "0.1.0"
description = "A python backend for Slack app that helps postting image for you."
authors = ["kuratowski <tonyshin@kaist.ac.kr>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^2.1"
django-allauth = "^0.39.1"
Pillow = "^5.4"
celery = "^4.2"

[tool.poetry.dev-dependencies]
selenium = "^3.141"
coverage = "^4.5"
fabric = "^2.4"
black = {version = "^18.3-alpha.0",allows-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
